Obey power budget
The NodeMCU board uses an LDO with a specified maximum output current of
1000mA. The grand total is 1200mA, but a) only for peaks and b) at this
point it starts damaging the device. The ESP8266 needs 100mA, probably a
lot less, since Wifi is disabled and the device idles most of the time.
The MAX4466 needs 0.02mA, which can be safely ignored when looking at
the power budgets.

However, one NeoPixel takes 50mA max, So the total worst case with all
pixels lit is 3200mA. Since not all pixels are white all of the time, it
is safe to assume that 1/3 of the power usage will do nicely. This is
accomplished by reducing the brightness accordingly.

This also lets the device run cooler, which is good for your skin.