Overhaul README
Images! Images for all electronics hardware components, a small demo how
the final product looks, and an image for the assembled (and fixed
components) board.

Also add more text on why to choose the NodeMCU ESP8266 board as base.

Be more verbose on how to assemble the thing and how to install the
software.

Add an F.A.Q.

And last but not least, add some warnings about putting electronics near
the face.