Switch to hardware-based PRNG
The software PRNG gets initialized by reading the analog pin A0. This
only works in a useful way when A0 is floating. This is not the case
here. So the software PRNG's random seed is of poor quality.

Furhtermore, the ESP8266 has a hardware random number generator. The
Arduino framework uses it when the random seed is not initialized.

Use the onboard hardware PRNG.