Add a Larson scanner animation
K.I.T.T. has it. Now we have it, too.

This Larson Scanner is a computational animation that uses a random base
color (not just red, like you-know-who). The color tries to be
outstanding from the white mouth. It scans over the closed mouth once.

Unlike the original, this misses reverse mode. The animation only runs
from left to right.