Fix up how animations display
The animation task stops the default display update task and emulates
it to display the animation itself. This is a hack, which even delay()'s
for the period of the display update task.

Shift responsibilities now. The animation task only sets which animation
to display and the display update task shows it. That makes it more
clear who does what and it avoids stopping and restarting tasks.

This has the benefit of better integration into existing debug print
infrastructure. The downside is lost locality and more global variables.