Introduce mouth effects
The mouth is shown in a boring constant white tone. Spice it up with
custom effects. An effect gets chosen and then applied to every mouth
frame.

The interface is chosen in a way that we can easily apply multiple
effects later.

This adds a bit more color and variation to Ze Voice Mask.