Add distort effect
This is a distortion of the mouth. The effect increases with the
loudness of the voice. Low voices (i.e. the idle mouth) will not likely
be distorted at all. Loud voices will almost always be completely
distorted.

The effect does not change pixel colors. It only shifts pixels up/down
or left/right.