More accurate Running_avg
Overhaul the Running_avg, as it has a lot of issues.

1) It does not work with negative integer values.
   The Rounding_adjustment is just plain wrong then.

2) With big SAMPLE sizes it does not approach the real average.
   E.g. when SAMPLE size is 1000 and the current average is 999,
   adding another host of 1000's does not make the jump over.
   The Rounding_adjustment does not provide enough room to update
   the average value by .1 or .2 or something simillarly small.
   It will always only be reduced to the rounded full integer.

3) It lacks priming.
   Together with 2, it might be possible that the average value
   never chnges.

Convert the internal representation to float. This gains accuracy and
makes the average slope up/down nicely. The external interface remains a
chosen data type that can be converted to/from float.

This also enablesi the use of negative values.

Finally, add priming. So even when initializing the average way off,
it converges to the real value rather quick.

The interface remains backwards-compatible. The average can still be
used like an ordinary integer type. When converting the internal float
to an integer type, rounding applies.

Add more tests to for the issues identified above.