Big MAX4466 overhaul
The analog signal from the MAX4466 jitters gallore. So this change
softens the analog signal by reading it 16 times in a row. There is even
an SDK function on the ESP8266 for that. To save memory, the 16 read
operations are split into 4x4. That only uses 8 slots, instead of 16.
That makes it kind of a double-smoothing algorithm.

The MAX4466 has a DC offset at 1024/2 for the zero-value, i.e. it is
totally silent. Unfortunately the offset varies on different hardware
modules from the real center by a considerable amount. This is bad,
since we want to detect silent signals, too. On some modules the offset
is off by sd much as 8%. With these modules the threashold to move the
mouth is hit quickly. This offset is now computed dynamically. It uses
the results from each of the split-up rounds in the double-smoothing
algorithm. If all values are the same, we can safely assume it is silent
and we have found the real center. Then steadily slope towards the
average of the found center points.

Since this module uses the Running_avg now, suck it into the project and
do not maintain it as a separate library.

Finally, make HardwareMAX4466 a Printable object. This allows the debug
printer to have better insights into the module's state.