Thin out multidimensional array
The array_P class was meant to hold arbitrary data as multidimensional
array in PROGMEM.  This works in clang, but unfortunately gcc ignores
the PROGMEM attribute. In gcc, the PROGMEM attribute needs to be applied
to a globally visible storage type. That means each variable has to be
declared with the decorator individually.

The variadic template boils down to a simple C-like multi-dimensional
array due to the use of value_type right from the start of the
recursion. Therfore, overloading operator[] does not work as the return
type is the underlying value_type.

Thin out the whole implementation. It's a simple poor-man's clone of
boost::multi_array now. Due to that, rename the whole thing.

This patch does not change anything in the final firmware image.