Move BwFrames to PROGMEM
BwFrames can be moved to PROGMEM easily. The underlying data type is
constant-iniatializable with the aggregate-initializer-lists used.

Mark all animation data as const-expressions. This enforces that the
aggregate-initializer performs static initialization. When someone
butchers it in the future, the compiler will refuse it.

Make the use of PROGMEM explicit in animation code, so users remember to
read it from flash. Functions that can handle frames from PROGMEM
themselves get a PGM_ prefix.