Move ColorFrames to PROGMEM
Moving ColorFrames to PROGMEM is a bit more involved. The basic data
type is RgbColor, which is not constant-initializable. that means the
exiting aggregate-initializations are performed at runtime. Moving these
variables to PROGMEM results in write attempts on read-only (PROGMEM)
memory.

Introduce PGM_RgbColor as a clever hack around that. It has the same
binary layout as RgbColor, but can be used in constant aggregate-
initializer-lists. It also silently converts to RgbColor.

Also mark all animation data as const-expressions. This enforces that
the aggregate-initializer performs static initialization. When someone
butchers it in the future, the compiler will refuse it.

Make the use of PROGMEM explicit in animation code, so users remembe to
read it from flash. Functions that can handle frames from PROGMEM
themselves get a PGM_ prefix.