effects/blend: Fix small edge case
The initialization of the blend effect tries to generate a dominant
color to blend to by filtering out one or more of the RGB base colors.

There is a small edge case where the blend effect essentially does not
do anything at all when all base colors are filtered out. Detect this
and create a grey base color instead.

Since all base colors need to be filtered out, hitting this edge case is
rather unlikely in practice.