Fix release make target
A simple

 $ make relase

builds with VOICEMASK_DEBUG defined. That is because the initial
invocation of make defines PLATFORMIO_BUILD_FLAGS and exports it to
nested invocations of make. The release target then invokes make again,
with PLATFORMIO_BUILD_FLAGS already defined. This way we end up with a
debug build.

Do not export PLATFORMIO_BUILD_FLAGS, pass it locally on each invocation
of the platformio command. Thus a nested make does not inherit its
previous definition.