README: Rework power supply
Remove the 9V block battery part. Nobody has built one with it. And it
is clearly more work than using a USB powerbank. Everyone seems to have
one lying around.

Add some hints on what USB cable to use. Generally a thinner cable is
more flexible and not too heavy around the neck. With strain on the USB
connector, it always breaks the male side on the cables. A magnetic
connector has two advantages. If the strain gets too much, it simply
disconnects. And to connect it in the first place, it just snaps in
place. No need to fiddle around on the micro-USB port.