build: Define default target in pio configuration
The current build system uses GNU make as entry point. However
issues pio commands directly. That makes builds and uploads run for all
environments, not only for nodemcuv2. This leads to spurious error
messages, especially for the nonexisting upload target in the native
environment.

Turn handling for the default environment upside down. Define it in the
platformio.ini file and extract it to be usable in GNU make. This way,
manual invocations of the pio command (e.g. through IDEs) get to know
the default environment automatically.

This does not change anything for GNU make as entry to the build system.

Note that support for running tests remains supported only from GNU make
at the moment.